+++ supergoon_engine/src/supergoon_engine/components/rigidbody_component.hpp
@@ -25,7 +25,7 @@ namespace Components
         void ApplyVelocity(const Gametime &);
         void ApplyVelocityByStepSolidsX(double step);
         void ApplyVelocityByStepSolidsY(double step);
-        bool TryAllMovementSteps(double full_step, double minimum_step, double& location_to_alter, double& velocity_to_alter, bool x_step);
+        bool TryAllMovementSteps(double full_step, double minimum_step, float& location_to_alter, float& velocity_to_alter, bool x_step);
         bool TryMovementStep(SDL_FRect& rect_to_check);
         void ApplyForce(Vector2 force, Vector2 constraint = Vector2());
         Vector2 velocity;
